Make `exitt` pass a return code
This makes it easier for other programs to detect whether NekCEM has
failed or succeeded. The immediate application is making it easier for
the test runner to tell whether a test has passed or not--before it
could be tricked by e.g. an early bailout due to incorrect
parameters.

See gh-81.