Fix final compiler warning in `cem_maxwell.F`
The warning was caused by a loop that could have caused an
out-of-bounds array access if `if3d` was true but `ldim` was set to
2. Of course, we know that this can't happen, but the compiler has no
way of determining this. This points out the more fundamental issue
that `if3d` should be declared a parameter--this will allow the
compiler to trim unnecessary 2D/3D branches during compilation.